refactor: move repetitive logic to separate components
