+++ src/pages/EmployeesCVsPage/CvsList/CvsList.styles.ts
@@ -1,0 +1,13 @@
+import { MenuItem, MenuList, styled } from '@mui/material';
+
+export const ItemMenu = styled(MenuItem)(() => ({
+  whiteSpace: 'pre-wrap',
+  wordWrap: 'break-word',
+  wordBreak: 'break-word',
+}));
+
+export const ListMenu = styled(MenuList)(() => ({
+  display: 'flex',
+  flexDirection: 'column',
+  rowGap: 18,
+}));
