+++ src/pages/EmployeesCVsPage/CvsList/CvsList.types.ts
@@ -1,0 +1,6 @@
+import { ICv } from '../../../interfaces/ICv.interface';
+
+export interface ICvsListProps {
+  data: ICv[];
+  onClick: (id: string) => () => void;
+}
