+++ src/pages/EmployeesCVsPage/CvsMenu/CvsMenu.styles.ts
@@ -1,4 +1,4 @@
-import { Card, MenuList, styled, Toolbar } from '@mui/material';
+import { Card, styled, Toolbar } from '@mui/material';
 
 export const CardMenu = styled(Card)(({ theme }) => ({
   width: '500px',
@@ -21,10 +21,4 @@ export const ToolbarMenu = styled(Toolbar)(({ theme }) => ({
   alignItems: 'center',
   width: '100%',
   backgroundColor: theme.palette.primary.main,
-}));
-
-export const ListMenu = styled(MenuList)(() => ({
-  display: 'flex',
-  flexDirection: 'column',
-  rowGap: 18,
 }));
