+++ src/pages/EmployeesCVsPage/CvsMenu/CvsMenu.tsx
@@ -1,7 +1,7 @@
 import { Drawer, IconButton } from '@mui/material';
 import React, { FC } from 'react';
 import CloseIcon from '@mui/icons-material/Close';
-import { ItemMenu, ListMenu } from '../EmployeesCVsPage.styles';
+import { CvsList } from '../CvsList/CvsList';
 import { ICvsMenuProps } from './CvsMenu.types';
 import * as Styled from './CvsMenu.styles';
 
@@ -32,13 +32,7 @@ export const CvsMenu: FC<ICvsMenuProps> = ({ data, open, showCvData, onClose }) 
             <CloseIcon />
           </IconButton>
         </Styled.ToolbarMenu>
-        <ListMenu>
-          {data?.user?.cvs?.map(({ id, name }) => (
-            <ItemMenu key={id} onClick={setCvData(id)}>
-              {name}
-            </ItemMenu>
-          ))}
-        </ListMenu>
+        <CvsList data={data?.user?.cvs || []} onClick={setCvData} />
       </Styled.CardMenu>
     </Drawer>
   );
