+++ src/pages/EmployeesCVsPage/Row/Row.types.ts
@@ -1,0 +1,4 @@
+export interface IRowProps {
+  title: string;
+  content: string;
+}
